Move `ArtifactRequest` class into its own file
